% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newick.R
\name{myProgram}
\alias{myProgram}
\title{A Cat Function}
\usage{
myProgram(x)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
