% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newick.R
\name{NewickReader}
\alias{NewickReader}
\title{A Newick-reader}
\usage{
NewickReader(x)
}
\arguments{
\item{x}{char}
}
\description{
This function allows you to read Newick-Strings
}
\examples{
((A,B),C)
Newick_reader()
}
