% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newick.R
\name{NewickReader}
\alias{NewickReader}
\title{Newick-reader}
\usage{
NewickReader(x, table = TRUE)
}
\arguments{
\item{x}{Dataset, like "((A,B),(((C,(D,E)),F),G),H)"}

\item{table}{Parameter for the table with the levels to each letter}
}
\description{
Function to read Newick strings
}
\examples{
NewickReader("((A,B),C)")
}
