% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newick.R
\name{NewickReader}
\alias{NewickReader}
\title{NewickReader}
\usage{
NewickReader(x)
}
\arguments{
\item{String}{A Newick-String}
}
\description{
This function allows you to read Newick-Strings.
}
\examples{
cat_function()
}
\keyword{cats}
