% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newick.R
\name{NewickReader}
\alias{NewickReader}
\title{A Newick-reader}
\usage{
NewickReader(x, table = TRUE)
}
\arguments{
\item{x}{char}

\item{table}{give back table}
}
\description{
This function allows you to read Newick-Strings
}
\examples{
NewickReader("((A,B),C)")
}
